% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StFecha.R
\name{formatFecha}
\alias{formatFecha}
\title{Title formatFecha}
\usage{
formatFecha(dtFecha)
}
\arguments{
\item{dtFecha}{is a String de fecha en formato yyyy-mm-dd}
}
\value{
dtFecha is a String de fecha en formato yyyymmdd
}
\description{
devuelve la fecha en forma yyyymmdd
}
\examples{
f1 <- c("2018-03-30")
f2 <- formatFecha(f1)

}
